Order fixes

+++ woocommerce-core-functions.php
@@ -27,7 +27,7 @@ function woocommerce_mail( $to, $subject, $message, $headers = "Content-Type: te
  **/
 if (!function_exists('woocommerce_get_page_id')) {
 	function woocommerce_get_page_id( $page ) {
-		return apply_filters('woocommerce_get_' . $page . '_page_id', get_option('woocommerce_' . $page . '_page_id'));
+		return (int) apply_filters('woocommerce_get_' . $page . '_page_id', get_option('woocommerce_' . $page . '_page_id'));
 	}
 }
